Change size_t in SerializedData to unsigned long

+++ satviz-consumer-native/include/satviz/Bindings.h
@@ -22,7 +22,7 @@ typedef struct CHeatUpdate {
 
 typedef struct SerializedData {
   const char *data;
-  size_t n;
+  unsigned long n;
 } SerializedData;
 
 // Graph
